+++ src/spyglass/spikesorting/v0/spikesorting_curation.py
@@ -324,7 +324,7 @@ class WaveformSelection(SpyglassMixin, dj.Manual):
 
 @schema
 class Waveforms(SpyglassMixin, dj.Computed):
-    use_transaction, _allow_insert = False, True
+    _use_transaction, _allow_insert = False, True
 
     definition = """
     -> WaveformSelection
@@ -525,7 +525,7 @@ class MetricSelection(SpyglassMixin, dj.Manual):
 
 @schema
 class QualityMetrics(SpyglassMixin, dj.Computed):
-    use_transaction, _allow_insert = False, True
+    _use_transaction, _allow_insert = False, True
 
     definition = """
     -> MetricSelection
@@ -963,6 +963,7 @@ class CuratedSpikeSorting(SpyglassMixin, dj.Computed):
     -> AnalysisNwbfile
     units_object_id: varchar(40)
     """
+    _use_transaction, _allow_insert = False, True
 
     class Unit(SpyglassMixin, dj.Part):
         definition = """