+++ src/spyglass/spikesorting/v0/spikesorting_sorting.py
@@ -11,6 +11,7 @@ import spikeinterface as si
 import spikeinterface.preprocessing as sip
 import spikeinterface.sorters as sis
 from spikeinterface.sortingcomponents.peak_detection import detect_peaks
+from tqdm import tqdm
 
 from spyglass.common.common_lab import LabMember, LabTeam
 from spyglass.settings import sorting_dir, temp_dir
@@ -248,20 +249,21 @@ class SpikeSorting(SpyglassMixin, dj.Computed):
         """Placeholder to override mixin method"""
         raise NotImplementedError
 
-    def cleanup(self):
-        """Clean up spike sorting directories that are not in the table.
+    def cleanup(self, dry_run=False):
+        """Clean up spike sorting directories that are not in the table."""
+        sort_dir = Path(sorting_dir)
+        tracked = set(self.fetch("sorting_path"))
+        all_dirs = {str(f) for f in sort_dir.iterdir() if f.is_dir()}
+        untracked = all_dirs - tracked
 
-        This should be run after AnalysisNwbFile().cleanup()
-        """
-        # get a list of the files in the spike sorting storage directory
-        dir_names = next(os.walk(sorting_dir))[1]
-        # now retrieve a list of the currently used analysis nwb files
-        analysis_file_names = self.fetch("analysis_file_name")
-        for dir in dir_names:
-            if dir not in analysis_file_names:
-                full_path = str(Path(sorting_dir) / dir)
-                logger.info(f"removing {full_path}")
-                shutil.rmtree(str(Path(sorting_dir) / dir))
+        if dry_run:
+            return untracked
+
+        for folder in tqdm(untracked, desc="Removing untracked folders"):
+            try:
+                shutil.rmtree(folder)
+            except PermissionError:
+                logger.warning(f"Permission denied: {folder}")
 
     @staticmethod
     def _get_sorting_name(key):